chore: move common helper to more common file
