+++ identity-wallet/src/state/did/validate_linked_verifiable_presentations.rs
@@ -1,6 +1,9 @@
 use crate::state::{
     core_utils::helpers::{download_logo, get_issuer_document, validate_credential_types},
-    did::validate_domain_linkage::{ValidationStatus, Verifier},
+    did::{
+        extract_url_from_did_web,
+        validate_domain_linkage::{ValidationStatus, Verifier},
+    },
 };
 use did_manager::Resolver;
 use futures::{
@@ -435,25 +438,6 @@ fn extract_logo_uri_from_display(display: &[Value]) -> Option<String> {
         .and_then(|logo| logo.get("uri").or(logo.get("url")))
         .and_then(|url| url.as_str())
         .map(ToString::to_string)
-}
-
-pub fn extract_url_from_did_web(did_web: &str) -> Option<Url> {
-    if let Some(did) = did_web.strip_prefix("did:web:") {
-        let url_str = if let Some(index_colon) = did.find(':') {
-            &did[..index_colon]
-        } else {
-            did
-        };
-
-        // TODO: quick hack to solve the percent-encoding issue in did:web:localhost%3A3033 (localhost:3033)
-        let url_decoded = url_str.replace("%3A", ":");
-
-        if let Ok(url) = Url::parse(&format!("http://{url_decoded}")) {
-            // TODO: the http:// hardcoded scheme is a hack to test with localhost
-            return Some(url);
-        }
-    }
-    None
 }
 
 #[cfg(not(feature = "test_utils"))]
